Backend/requirements.txt
django==4.2.1
django-rest-framework
djangorestframework
djangorestframework-simplejwt
django-extensions
pyyaml
requests
django-cors-headers
openai-whisper
setuptools-rust
sounddevice
wavio
pydub
transformers
numpy<2
python-dotenv
librosa
torchvision
daphne
channels
tensorflow
tf-keras
speechbrain
torchaudio
opencv-python
keras-facenet